probegin/requirements.txt
Django==2.2.10
psycopg2==2.8.4
pytest-django==3.8.0
